telegram_bot/requirements.txt
# Helios2 Telegram Bot Requirements
python-telegram-bot
httpx
python-dotenv
